+++ release-notes.md
@@ -14,6 +14,7 @@ Fixes:
   is included in the `UiOptions.StudyMainTags` configuration)
 - Fix the MediaStorageSOPClassUID of STL files that are added to an existing study.
 - Stop logging the HTTP headers when redirecting from root (https://discourse.orthanc-server.org/t/orthanc-explorer-2-error-logs-upon-redirection-to-new-ui/6232)
+- Fix issue when changing only the casing of the PatientID (https://discourse.orthanc-server.org/t/possible-bug-when-changing-casings-in-patientid-using-explorer2/6124)
 
 
 1.9.0 (2025-08-13)